Align MessageFeedbackEntity/SourceFeedbackEntity id and DTOs to Long

+++ src/main/java/zas/admin/zec/backend/actions/analyze/FeedbackDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class FeedbackDTO {
     // --- Message feedback ---
     public record MessageFeedback(
-            Integer id,
+            Long id,
             String userUuid,
             String conversationUuid,
             String messageUuid,
@@ -28,7 +28,7 @@ public class FeedbackDTO {
 
     // --- Source feedback row ---
     public record SourceFeedback(
-            Integer id,
+            Long id,
             String userUuid,
             String conversationUuid,
             String messageUuid,
